requirements.txt
# Core ML
torch>=2.1.0,<2.5.0
numpy>=1.24.0,<2.0.0

# Data
datasets>=2.14.0
tokenizers>=0.15.0

# Config and experiments
pyyaml>=6.0
tqdm>=4.66.0

# Analysis and plotting
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Testing and dev
pytest>=7.4.0
pytest-cov>=4.1.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
